rename pkg to core
